Fix two failing unit tests
The first failure was something shifting in the invoke API. I didn't
fix it directly, but instead brought in the dedicated `doctrine`
library I made for this purpose that I know works.

The second one seems to be small adjustments due to graphviz changing.
Seems reasonable to adjust the fixture for now (maybe I can make it
more tolerant later).